requirements.txt
future==0.18.2
iso8601==0.1.13
pynput==1.7.1
PyQt5==5.15.2
PyQt5-sip==12.8.1
pyserial==3.4
PySide2==5.15.2
PyYAML==5.3.1
shiboken2==5.15.2
six==1.15.0
